troubleshooting and r_engineering

+++ app/controllers/wikis_controller.rb
@@ -4,32 +4,33 @@ class WikisController < ApplicationController
   def index
     @wikis = WikiPolicy::Scope.new(current_user, Wiki).resolve
   end
-#
+ #
   def show
     @wiki = Wiki.find(params[:id])
-    if @wiki.private?
-      if @wiki.user == current_user
-        wiki_path
-      elsif current_user.standard?
-        flash[:alert] = "You must be a premium user to view private wikis."
+    if current_user.present? # setting collaborations
+      colaborators = []
+      @wiki.collaborators.each do |collaborator| # if collaborator add to collaborators array
+        collaborators << collaborator.email
+      end
+      unless (@wiki.private == false) || @wiki.user == current_user || collaborators.include(current_user.email) || current_user.admin?
+        flash[:alert] = "You do not have the proper authorization to view private wikis."
         redirect_to new_charge_path
-      elsif @wiki.users.exclude?(current_user)
+      end
+    else
         flash[:alert] = "You must be a collaborator to view that wiki."
         redirect_to wikis_path
       end
-    else
-      wiki_path
-    end
   end
 #
   def new
     @wiki = Wiki.new
     authorize @wiki
   end  
-
+#
   def create
     @wiki = Wiki.new(wiki_params)
     @wiki.user = current_user
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Your Wiki was successfully saved."
@@ -39,17 +40,20 @@ class WikisController < ApplicationController
       render :new
     end 
   end
-
+#
   def edit
     @wiki = Wiki.find(params[:id])
     @user = User.all
+    authorize @wiki
   end
 #
   def update
     @wiki = Wiki.find(params[:id])
     @wiki.assign_attributes(wiki_params)
+    authorize @wiki
 
-    if @wiki.save
+    if @wiki.save && (@wiki.user == current_user || current_user.admin?)
+      @wiki.collaborators = Collaborator.update_collaborators(params[:wiki][:collaborators])
       flash[:notice] = "Your wiki was successfully updated."
       redirect_to @wiki
     else
@@ -57,9 +61,11 @@ class WikisController < ApplicationController
       render :edit
     end  
   end
-
+#
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
+
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
       redirect_to @wiki
